+++ packages/blocks-library/icon/src/Block.php
@@ -96,17 +96,12 @@ class Block implements BlockInterface {
 			$this->args['plugin_version']
 		);
 
-		// Remove all img tags.
-		foreach ($dom->find('img') as $img) {
-			$img->outertext = '';
-		}
-
 		$value = $block['attrs']['blockeraIcon']['value'] ?? $this->fallback_value;
 
-		$figure = $dom->findOne('figure');
+		$figure = $dom->findOne('img');
 
 		if ($figure) {
-			$figure->innerhtml = sprintf(
+			$figure->outerhtml = sprintf(
 				'<span title="%1$s" role="img" class="wp-block-image__icon">%2$s</span>', 
 				sprintf( 
 					// translators: %s is the icon name.
@@ -114,16 +109,11 @@ class Block implements BlockInterface {
 					str_replace('-', ' ', $value['icon'] ?? '') 
 				), 
 				$feature->getIconHTML($value)
-			) . $figure->innerhtml;
+			);
 
 			$svg = $dom->findOne('svg');
 
 			if ($svg) {
-				$size  = $block['attrs']['blockeraIconSize']['value'] ?? '1.33em';
-				$color = $block['attrs']['blockeraIconColor']['value'] ?? 'currentColor';
-
-				$style = 'width:' . $size . ';height:' . $size . ';fill:' . $color . ';color:' . $color . ';';
-
 				$transform = '';
 
 				// Handle icon rotate.
@@ -145,11 +135,17 @@ class Block implements BlockInterface {
 				}
 
 				if (! empty($transform)) {
-					$style .= 'transform:' . $transform . ';';
-				}
-
-				$svg->style = $style;
+					$transform .= 'transform:' . $transform . ';';
+					$svg->setAttribute('style', $transform);
+				} else {
+					$svg->removeAttribute('style');
+				}				
+			} else {
+				$svg->removeAttribute('style');
 			}
+
+			$svg->setAttribute('focusable', 'false');
+			$svg->setAttribute('aria-hidden', 'true');
 		}
 
 		return $dom->outerhtml;