+++ packages/blocks-core/CHANGELOG.md
@@ -4,11 +4,18 @@
 - Added a way for blocks to get a unique ID automatically, making them work more smoothly.
 - Added support for blocks to better integrate with the block editor's features.
 
+### Improvements
+- `List` block: `Icon` inner block added.
+- `List Item` block: `Icon` inner block added.
+- `Button` block: `Icon` inner block added.
+
+
 ### Automated Tests
 - Added E2E tests to check icon functionality of `Categories` block.
 - Added E2E tests to check icon functionality of `Page List` block.
 - Added E2E tests to check icon functionality of `Home Link` block.
 - Added E2E tests to check icon functionality of `Navigation Link` block.
+- Added E2E tests to check icon functionality of `Navigation Submenu` block.
 - Added E2E tests to check icon functionality of `Navigation Submenu` block.
 
 ## 1.4.0 (2025-05-19)